Refactor Editable Cell
