--- front/src/modules/ui/components/editable-cell/CellBaseContainer.tsx
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-export const CellBaseContainer = styled.div`
-  position: relative;
-  box-sizing: border-box;
-  height: 32px;
-  display: flex;
-  align-items: center;
-  width: 100%;
-  cursor: pointer;
-  user-select: none;
-`;
