--- front/src/modules/ui/components/editable-cell/CellNormalModeContainer.tsx
@@ -1,12 +1,0 @@
-import styled from '@emotion/styled';
-
-export const CellNormalModeContainer = styled.div`
-  display: flex;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  overflow: hidden;
-
-  padding-left: ${(props) => props.theme.spacing(2)};
-  padding-right: ${(props) => props.theme.spacing(2)};
-`;
