--- front/src/modules/ui/components/editable-cell/EditableCellMenu.tsx
@@ -1,67 +1,0 @@
-import { ReactElement } from 'react';
-import styled from '@emotion/styled';
-import { useRecoilState } from 'recoil';
-
-import { isSomeInputInEditModeState } from '../../tables/states/isSomeInputInEditModeState';
-
-import { CellBaseContainer } from './CellBaseContainer';
-import { EditableCellMenuEditMode } from './EditableCellMenuEditMode';
-
-const EditableCellMenuNormalModeContainer = styled.div`
-  display: flex;
-  align-items: center;
-  width: calc(100% - ${(props) => props.theme.spacing(5)});
-  height: 100%;
-  overflow: hidden;
-`;
-
-type OwnProps = {
-  editModeContent: ReactElement;
-  nonEditModeContent: ReactElement;
-  editModeHorizontalAlign?: 'left' | 'right';
-  editModeVerticalPosition?: 'over' | 'below';
-  isEditMode?: boolean;
-  isCreateMode?: boolean;
-  onOutsideClick?: () => void;
-  onInsideClick?: () => void;
-};
-
-// TODO: refactor
-export function EditableCellMenu({
-  editModeContent,
-  nonEditModeContent,
-  editModeHorizontalAlign = 'left',
-  editModeVerticalPosition = 'over',
-  isEditMode = false,
-  onOutsideClick,
-  onInsideClick,
-}: OwnProps) {
-  const [isSomeInputInEditMode, setIsSomeInputInEditMode] = useRecoilState(
-    isSomeInputInEditModeState,
-  );
-
-  function handleOnClick() {
-    if (!isSomeInputInEditMode) {
-      onInsideClick?.();
-      setIsSomeInputInEditMode(true);
-    }
-  }
-
-  return (
-    <CellBaseContainer onClick={handleOnClick}>
-      <EditableCellMenuNormalModeContainer>
-        {nonEditModeContent}
-      </EditableCellMenuNormalModeContainer>
-      {isEditMode && (
-        <EditableCellMenuEditMode
-          editModeContent={editModeContent}
-          editModeHorizontalAlign={editModeHorizontalAlign}
-          editModeVerticalPosition={editModeVerticalPosition}
-          isEditMode={isEditMode}
-          onOutsideClick={onOutsideClick}
-          onInsideClick={onInsideClick}
-        />
-      )}
-    </CellBaseContainer>
-  );
-}
