+++ front/src/modules/ui/components/editable-cell/types/EditableRelationCreateButton.tsx
@@ -16,9 +16,4 @@ export const EditableRelationCreateButton = styled.button`
   height: 31px;
   background: none;
   gap: 8px;
-  // :hover {
-  //   background: rgba(0, 0, 0, 0.04);
-  //   color: ${(props) => props.theme.text100};
-  // }
-  // margin-bottom: calc(${(props) => props.theme.spacing(1)} / 2);
 `;
